requirements.txt
asgiref==3.8.1
Django==5.0.3
django-monthfield==0.1.4
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
